Fix Main

+++ main.py
@@ -4,7 +4,7 @@ import requests
 import logging
 
 from dotenv import load_dotenv
-print('Varna')
+
 if os.path.exists('log.txt'):
     with open('log.txt', 'r+') as f:
         f.truncate(0)
@@ -37,16 +37,17 @@ try:
     if len(UPSTREAM_REPO) == 0:
        raise TypeError
 except TypeError:
-    UPSTREAM_REPO = "https://github.com/VarnaX/EvaMaria"
+    UPSTREAM_REPO = None
 
-if os.path.exists('.git'):
-    subprocess.run(["rm", "-rf", ".git"])
+if UPSTREAM_REPO is not None:
+    if os.path.exists('.git'):
+        subprocess.run(["rm", "-rf", ".git"])
 
-subprocess.run([f"git init -q \
-                  && git config --global user.email e.anastayyar@gmail.com \
-                  && git config --global user.name mltb \
-                  && git add . \
-                  && git commit -sm update -q \
-                  && git remote add origin {UPSTREAM_REPO} \
-                  && git fetch origin -q \
-                  && git reset --hard origin/master -q"], shell=True)
+    subprocess.run([f"git init -q \
+                      && git config --global user.email e.anastayyar@gmail.com \
+                      && git config --global user.name mltb \
+                      && git add . \
+                      && git commit -sm update -q \
+                      && git remote add origin {UPSTREAM_REPO} \
+                      && git fetch origin -q \
+                      && git reset --hard origin/master -q"], shell=True)
